requirements.txt
asgiref==3.9.0
Django==5.2.4
django_components==0.141.2
djangorestframework==3.16.0
djc_core_html_parser==1.0.2
sqlparse==0.5.3
typing_extensions==4.14.1
